Initial commit for change project structe, separated Unit Test from Integration Test folders
