+++ Redis/src/main/java/com/dantebasso/study/redis/RedisStudyApplication.java
@@ -1,4 +1,4 @@
-package com.redis.study.RedisStudy;
+package com.dantebasso.study.redis;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
