+++ Redis/src/test/java/com/dantebasso/study/redis/service/CarServiceTest.java
@@ -1,0 +1,17 @@
+package com.dantebasso.study.redis.service;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+/**
+ * @author Dante Basso <${company.email}>
+ * @since 16-06-2022
+ */
+public class CarServiceTest {
+
+    @Test
+    public void test() {
+        Assertions.assertTrue(Boolean.TRUE);
+    }
+
+}
